Add "shortcut methods" for reducing boilerplate code
Previously a bit of boilerplate code was involved when adding simple states and transitions. To improve this I've overloaded the normal AddState, AddTransition, ... methods with new methods that take the normal parameters of State and Transition directly. This means that the user no longer has to pass a state instance in, but instead, the fsm creates one under the hood.