Add getters for states and transitions
- Add methods to get all states and all transitions, trigger transitions, transitions from any, trigger transitions from any

This change has been made in order to make the members marked with the "internal" access modifier "private" again. They needed to be "internal" for the current implementation of the animator debug graph generation (see EditorStateMachineShortcuts.cs). However this exposed implementation details to the outside (at least within the assembly).

Warning: The code currently does not compile